requirements.txt
django>=4.1.2,<4.2
uwsgi>=2.0.20,<2.1
user-agents==2.2.0
maxminddb>=2.2.0,<2.3
djangorestframework>=3.13.1,<3.14
drf-spectacular>=0.24.2,<0.25
django-allauth>=0.51.0,<0.52.0
psycopg2>=2.9.3,<2.10
django-ses>=3.1.0,<3.2
dj_rest_auth>=2.2.5,<2.3
django-debug-toolbar>=3.7.0,<3.8.0
